please add data directory / link to the resources
I have been trying to work on this lesson. Everything is explained perfectly in a clear way. However, students are learning more by doing the exercise as they follow the instruction.  The only minor issue here is that I could not find the right link to the data resource of this lesson.
From the following code, I would guess there would be a data directory and a few images in it. 
 image = skimage.io.imread(fname="data/eight.tif")
That would be awesome to provide the right link to the data directory.